Move DI control-plane scripts into util/di/, strip comments, inline single-use names
- Relocate di-control-plane.py + di-api-requests.json under util/di/.
- Remove the module docstring and the test.yml step comment.
- Rename render() -> create_di_request_body(), call() -> call_di().
- Inline single-use module constants (HERE/REQUESTS/REGION/API_URL/NOT_FOUND)
  and the set_github_env helper.
Behavior unchanged; all 4 verbs verified end-to-end from the new location.
